modularised unit tests and added more cases
